strategy/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.6
python-dotenv==1.0.0
pydantic==1.10.7
httpx==0.24.1
PyJWT==2.6.0